Rename files to be standard Python naming.
